Population Culling & Nuking
Population Culling
--------------------
Once population show signs of stagnation in fitness growth, a portion of the population will be replaced by new randomly generated replacements

The portion of the population is determined by either lowest fitness or uniqueness

Population Nuking
---------------------
After a certain amount of Culling (MaxCullCount) has been performed and stagnation is still present, the population with the exception of elites will be replaced entirely with new randomly generated agents.

However, nuking seems to be too drastic attempt to reintro diversity to the population. It almost entire push the training process back to Generation 0.