Revamp of GeneticTrainer (TrainingModel / Individual classes for Genetic Operators / Removal of Participants)
Revamp of Genetic Trainer
-----------------------------
a. GeneticTrainer has been quite inflexible when it comes to iterating due to reminents of std::couts and different genetic operators being reuse through comment and uncommenting.

b. A more friendly interface is needed to allow modification in each iteration

TrainingModel
----------------
A struct that contains the various properties that will be modify in each iteration (eg. Type of Selection/Mutation, Amount of Generations to be run)

Now, instead of passing multiple parameters, a pair of TrainingModel and FeedbackLayer will be pass to Scheduler to facilite the iteration

PS. FeedbackLayer is used to define the extent of how much info std::couts will be displaying (eg. Generational / Match)

Individual Classes for GeneticOperator
------------------------------------------
The execution of each Genetic Operator (Selection, Crossover & Mutation) will now be done in their individual class (Selector, Crossoverer & Mutator)

The specific approach of each operator will be dictate by the Method stated within the TrainingModel. All GeneticTrainer has to do is to call their respective SelectFrom/Cross/Mutate function.

More specific parameters to each operator that cannot be found in TrainingModel, can be found in these classes